Add separate component for simplicity
